requirements.txt
dnspython==2.7.0
Faker==37.3.0
greenlet==3.2.3
mysql-connector-python==9.3.0
numpy==2.2.6
pandas==2.2.3
psycopg2-binary==2.9.10
pymongo==4.13.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
